==> CODE/8.0.Linkage/OLD_CODE_pre2023/6.pull_annotation_GLM_LD_snps.R <==
### Annotations of the SNPs emanating from the GLM + LD analysis
### 


## load packages
library(tidyverse)
library(magrittr)
library(reshape2)
library(data.table)
library(SeqArray)
library(gdsfmt)
library(SNPRelate)
library(foreach)
library(doMC)
registerDoMC(10) ## using an i-job ##ijob -A jcbnunez -c 2 --mem=120G  --partition=largemem

## import gds
genofile <- seqOpen("/project/berglandlab/DEST_Charlottesville_TYS/DEST_pipeline_output/dest.all.PoolSNP.001.50.10Mar2021.ann.gds")
## import metdata from glm + ld analyses
glm_ld_outs <- fread("/scratch/yey2sn/Overwintering_ms/7.LD/GLM_LD_outliers.txt")

## create snpo guide file
snps.dt <- data.table(chr=seqGetData(genofile, "chromosome"),
                      pos=seqGetData(genofile, "position"),
                      variant.id=seqGetData(genofile, "variant.id"))

snps.dt %>%
  filter(chr == "2L") %>%
  filter(pos %in% glm_ld_outs$pos[which(is.na(glm_ld_outs$inversion_marker))]) ->
  snps.dt.out

## import the glm outlier + ld object
seqSetFilter(genofile, variant.id=snps.dt.out$variant.id)

subset_geno = genofile

### Process annotations using a list decomposition format
### I will use a foreach loop to extract the annotation of each outlier SNP
annotation_set <- foreach(i=1:dim(snps.dt.out)[1],
                          .combine = "rbind")%do%{ ## open DO loop ## MAYBE CHANGE TO DOPAR?
  
  print(i)
  subset_geno = genofile
  
  seqSetFilter(subset_geno, variant.id=snps.dt.out$variant.id[i])
  
      tmp_annot <-  data.frame(annot = seqGetData(subset_geno, "annotation/info/ANN")[[2]])
      
      #extract annotation and parition into categories
      tmp_annot %>%
        separate(annot,
                 into = c(
                   "Allele",
                   "Annotation",
                   "Putative_impact",
                   "Gene_Name",
                   "Gene_ID",
                   "Feature_type",
                   "Feature_ID",
                   "Transcript_biotype",
                   "Rank",
                   "HGVS.c",
                   "HGVS.p",
                   "cDNA_position",
                   "CDS_position",
                   "Protein_position",
                   "Distance_to_feature",
                   "Notes"),
                 sep = "\\|",
                 convert = T
                 ) -> tmp_annot_sep
      
      ## include annotaions as well as position and chromosome
      obj_to_rturn <- data.frame(chr = snps.dt.out$chr[i],
                                 pos = snps.dt.out$pos[i],
                                 tmp_annot_sep
                                  )
      
      return(obj_to_rturn) ##<---- return this object to the output object
      
} ## close do loop
#fix a boolean misinterpretation
annotation_set$Allele[which(annotation_set$Allele == "TRUE")] = "T"



### add priority scoring
### 
annotation_set %<>%
  mutate(priority_scoring = case_when(Annotation == "3_prime_UTR_variant" ~ 40,
                                      Annotation == "3_prime_UTR_premature_start_codon_gain_variant" ~ 45,
                                      Annotation == "5_prime_UTR_variant" ~ 40,
                                      Annotation == "downstream_gene_variant" ~ 0,
                                      Annotation == "intergenic_region" ~ 0,
                                      Annotation == "intron_variant" ~ 20,
                                      Annotation == "missense_variant" ~ 80,
                                      Annotation == "missense_variant&splice_region_variant" ~ 100,
                                      Annotation == "non_coding_transcript_exon_variant" ~ 15,
                                      Annotation == "splice_region_variant" ~ 70,
                                      Annotation == "splice_region_variant&intron_variant" ~ 70,
                                      Annotation == "synonymous_variant" ~ 50,
                                      Annotation == "upstream_gene_variant" ~ 0
                                      ))

## create unique annotation list based on priority_scoring
annotation_set %>% 
  group_by(chr, pos) %>%
  slice_max(priority_scoring, with_ties = F) ->
  annotation_set_prioritized

write.table(annotation_set_prioritized,
            file = "./GLM_LD_outliers_annotation_priorized.txt", 
            append = FALSE, 
            quote = FALSE, 
            sep = "\t",
            eol = "\n", 
            na = "NA", 
            dec = ".", 
            row.names = FALSE,
            col.names = TRUE, 
            qmethod = c("escape", "double"),
            fileEncoding = "")


### Analyze annotations
annotation_set %>% 
  filter(Annotation %in% c("missense_variant") ) %>%
  group_by(chr, pos, Annotation, Gene_Name, HGVS.c,	HGVS.p) %>%
  summarise(N= n(),
            Mean_priority = mean(priority_scoring) ) -> missense_variants_summary

write.table(missense_variants_summary,
            file = "./missense_variants.annot.txt", 
            append = FALSE, 
            quote = FALSE, 
            sep = "\t",
            eol = "\n", 
            na = "NA", 
            dec = ".", 
            row.names = FALSE,
            col.names = TRUE, 
            qmethod = c("escape", "double"),
            fileEncoding = "")

annotation_set %>% 
  filter(Annotation %in% c("missense_variant") ) %>%
  group_by(chr, pos, Annotation, Gene_Name, HGVS.c,	HGVS.p) %>%
  summarise(N= n()) %>%
  group_by(pos) %>%
  slice_head() -> missense_variants_nonRedundant

write.table(missense_variants_nonRedundant,
            file = "./missense_variants_nonRedundant.txt", 
            append = FALSE, 
            quote = FALSE, 
            sep = "\t",
            eol = "\n", 
            na = "NA", 
            dec = ".", 
            row.names = FALSE,
            col.names = TRUE, 
            qmethod = c("escape", "double"),
            fileEncoding = "")
######
annotation_set_prioritized <- fread("../7.LD/GLM_LD_outliers_annotation_priorized.txt")

annotation_set_prioritized %>%
  .[grep("CG", .$Gene_Name, invert =  T),] %>%
  filter(Putative_impact %in% c("MODERATE")) %>%
  group_by(Gene_Name) %>%
  slice_head() %>%
  group_by(Gene_Name) %>%
  summarise(Pos_mean = mean(pos)) %>%
  arrange(Pos_mean) %>%
  mutate(vert_pos = as.numeric(rownames(.))) %>%
  ggplot(aes(
    x=Pos_mean,
    y=vert_pos,
    label = Gene_Name
  )) +
  geom_text(size = 4) -> gene_labels
ggsave(gene_labels, file = "gene_labels.pdf",
       width = 9,
       height = 5,)
